Renamed collectors to sources
