+++ src/routes/(sidebar)/crud/sources/+page.svelte
@@ -4,8 +4,8 @@
 	import { TableHeadCell, Toolbar, ToolbarButton } from 'flowbite-svelte';
 	import { CogSolid, DotsVerticalSolid, EditOutline } from 'flowbite-svelte-icons';
 	import { ExclamationCircleSolid, TrashBinSolid } from 'flowbite-svelte-icons';
-	import Collector from './Collectors.svelte';
-	import CollectorsList from './CollectorsList.svelte';
+	import Source from './Sources.svelte';
+	import SourcesList from './SourcesList.svelte';
 	import GCSForm from './GCS.svelte';
 	import AzureForm from './Azure.svelte';
 	import DriveForm from './Drive.svelte';
@@ -21,11 +21,11 @@
 	import MetaTag from '../../../utils/MetaTag.svelte';
 
 	let hidden: boolean = true; // modal control
-	let drawerComponent: ComponentType = Collector; // drawer component
+	let drawerComponent: ComponentType = Source; // drawer component
 
 	const toggle = (component: typeof SvelteComponent | undefined, config: any) => {
 		if (!component) {
-			console.error('No component found for this collector type:');
+			console.error('No component found for this Source type:');
 			return; // Handle this case appropriately, maybe show an error message
 		}
 		drawerComponent = component;
@@ -34,14 +34,14 @@
 	};
 
 	function toggleList() {
-		drawerComponent = CollectorsList;
+		drawerComponent = SourcesList;
 		hidden = !hidden;
 	}
 	let configuration = {};
-	type CollectorComponents = {
+	type SourceComponents = {
 		[key: string]: any;
 	};
-	const nameToCollector: CollectorComponents = {
+	const nameToSource: SourceComponents = {
 		'Google Cloud Storage': GCSForm,
 		Azure: AzureForm,
 		'Google Drive': DriveForm,
@@ -55,28 +55,28 @@
 		'Local Storage': LocalStorageForm
 	};
 
-	const path: string = '/crud/collectors';
-	const description: string = 'Collectors examaple - Querent Admin Dashboard';
-	const title: string = 'Querent Admin Dashboard - Collectors';
-	const subtitle: string = 'Collectors';
-	let collectors_list: any[] = [];
-	function handleCollectorSaved(event: {
+	const path: string = '/crud/source';
+	const description: string = 'Sources examaple - Querent Admin Dashboard';
+	const title: string = 'Querent Admin Dashboard - Sources';
+	const subtitle: string = 'Sources';
+	let sources_list: any[] = [];
+	function handleSourceSaved(event: {
 		detail: { name: any; technology: any; description: any; configuration: any };
 	}) {
 		const { name, technology, description, configuration } = event.detail;
-		const index = collectors_list.findIndex((collector) => collector.name === name);
+		const index = sources_list.findIndex((source) => source.name === name);
 
 		if (index !== -1) {
-			collectors_list[index] = { name, technology, description, configuration };
-			collectors_list = [...collectors_list];
+			sources_list[index] = { name, technology, description, configuration };
+			sources_list = [...sources_list];
 		} else {
-			collectors_list.push({ name, technology, description, configuration });
-			collectors_list = [...collectors_list];
+			sources_list.push({ name, technology, description, configuration });
+			sources_list = [...sources_list];
 		}
 	}
 	function getComponent(name: string): typeof SvelteComponent | undefined {
 		console.log(name);
-		return nameToCollector[name as keyof typeof nameToCollector] || undefined;
+		return nameToSource[name as keyof typeof nameToSource] || undefined;
 	}
 </script>
 
@@ -86,15 +86,15 @@
 	<div class="p-4">
 		<Breadcrumb class="mb-5">
 			<BreadcrumbItem home>Home</BreadcrumbItem>
-			<BreadcrumbItem href="/crud/collectors">Collectors</BreadcrumbItem>
-			<BreadcrumbItem>Collectors</BreadcrumbItem>
+			<BreadcrumbItem href="/crud/sources">Sources</BreadcrumbItem>
+			<BreadcrumbItem>Sources</BreadcrumbItem>
 		</Breadcrumb>
 		<Heading tag="h1" class="text-xl font-semibold text-gray-900 dark:text-white sm:text-2xl">
-			All collectors
+			All sources
 		</Heading>
 
 		<Toolbar embedded class="w-full py-4 text-gray-500 dark:text-gray-400">
-			<Input placeholder="Search for collectors" class="me-6 w-80 border xl:w-96" />
+			<Input placeholder="Search for sources" class="me-6 w-80 border xl:w-96" />
 			<ToolbarButton
 				color="dark"
 				class="m-0 rounded p-1 hover:bg-gray-100 focus:ring-0 dark:hover:bg-gray-700"
@@ -121,40 +121,40 @@
 			</ToolbarButton>
 
 			<div slot="end" class="space-x-2">
-				<Button class="whitespace-nowrap" on:click={toggleList}>Add new collector</Button>
+				<Button class="whitespace-nowrap" on:click={toggleList}>Add new source</Button>
 			</div>
 		</Toolbar>
 	</div>
 	<Table>
 		<TableHead class="border-y border-gray-200 bg-gray-100 dark:border-gray-700">
 			<TableHeadCell class="w-4 p-4"><Checkbox /></TableHeadCell>
-			{#each ['Collector', 'Company', 'Description'] as title}
+			{#each ['Source', 'Company', 'Description'] as title}
 				<TableHeadCell class="ps-4 font-normal">{title}</TableHeadCell>
 			{/each}
 			<TableHeadCell class="pe-100 ps-4 text-right font-normal">Actions</TableHeadCell>
 		</TableHead>
 		<TableBody>
-			{#each collectors_list as collector}
+			{#each sources_list as source}
 				<TableBodyRow class="text-base">
 					<TableBodyCell class="w-4 p-4"><Checkbox /></TableBodyCell>
 					<TableBodyCell class="flex items-center space-x-6 whitespace-nowrap p-4">
 						<div class="text-sm font-normal text-gray-500 dark:text-gray-400">
 							<div class="text-base font-semibold text-gray-900 dark:text-white">
-								{collector.name}
+								{source.name}
 							</div>
 						</div>
 					</TableBodyCell>
-					<TableBodyCell class="p-4">{collector.technology}</TableBodyCell>
+					<TableBodyCell class="p-4">{source.technology}</TableBodyCell>
 					<TableBodyCell
 						class="max-w-md overflow-hidden truncate p-4 text-base font-normal text-gray-500 dark:text-gray-400 xl:max-w-lg"
-						>{collector.description}</TableBodyCell
+						>{source.description}</TableBodyCell
 					>
 
 					<TableBodyCell class="flex justify-end space-x-2 p-4">
 						<Button
 							size="sm"
 							class="gap-2 px-3"
-							on:click={() => toggle(getComponent(collector.name), collector.configuration)}
+							on:click={() => toggle(getComponent(source.name), source.configuration)}
 						>
 							<EditOutline size="sm" /> Edit
 						</Button>
@@ -166,14 +166,14 @@
 </main>
 
 <Drawer placement="right" transitionType="fly" bind:hidden>
-	{#if drawerComponent === CollectorsList}
-		<CollectorsList bind:hidden bind:configuration on:collectorSaved={handleCollectorSaved} />
+	{#if drawerComponent === SourcesList}
+		<SourcesList bind:hidden bind:configuration on:sourceSaved={handleSourceSaved} />
 	{:else}
 		<svelte:component
 			this={drawerComponent}
 			bind:hidden
 			bind:configuration
-			on:collectorSaved={handleCollectorSaved}
+			on:sourceSaved={handleSourceSaved}
 		/>
 	{/if}
 </Drawer>
