+++ src/routes/(sidebar)/crud/sources/Delete.svelte
