+++ src/routes/(sidebar)/crud/users/+page.svelte
@@ -4,20 +4,20 @@
 	import { TableHeadCell, Toolbar, ToolbarButton } from 'flowbite-svelte';
 	import { CogSolid, DotsVerticalSolid, EditOutline } from 'flowbite-svelte-icons';
 	import { ExclamationCircleSolid, TrashBinSolid } from 'flowbite-svelte-icons';
-	import RealtimeCollectors from '../../../data/realtime_collectors.json';
-	import Collector from '../collectors/Collectors.svelte';
+	import RealtimeSources from '../../../data/realtime_sources.json';
+	import Source from '../sources/Sources.svelte';
 	import type { ComponentType } from 'svelte';
 	import MetaTag from '../../../utils/MetaTag.svelte';
 
 	let hidden: boolean = true; // modal control
-	let drawerComponent: ComponentType = Collector; // drawer component
+	let drawerComponent: ComponentType = Source; // drawer component
 
 	const toggle = (component: ComponentType) => {
 		drawerComponent = component;
 		hidden = !hidden;
 	};
 
-	const path: string = '/crud/collectors';
+	const path: string = '/crud/sources';
 	const description: string = 'CRUD products examaple - Querent Admin Dashboard';
 	const title: string = 'Querent Admin Dashboard - CRUD Products';
 	const subtitle: string = 'CRUD Products';
@@ -29,7 +29,7 @@
 	<div class="p-4">
 		<Breadcrumb class="mb-5">
 			<BreadcrumbItem home>Home</BreadcrumbItem>
-			<BreadcrumbItem href="/crud/collectors">E-commerce</BreadcrumbItem>
+			<BreadcrumbItem href="/crud/sources">E-commerce</BreadcrumbItem>
 			<BreadcrumbItem>Products</BreadcrumbItem>
 		</Breadcrumb>
 		<Heading tag="h1" class="text-xl font-semibold text-gray-900 dark:text-white sm:text-2xl">
@@ -73,24 +73,24 @@
 			<TableHeadCell class="pe-100 ps-4 text-right font-normal">Actions</TableHeadCell>
 		</TableHead>
 		<TableBody>
-			{#each RealtimeCollectors as collector}
+			{#each RealtimeSources as source}
 				<TableBodyRow class="text-base">
 					<TableBodyCell class="w-4 p-4"><Checkbox /></TableBodyCell>
 					<TableBodyCell class="flex items-center space-x-6 whitespace-nowrap p-4">
 						<div class="text-sm font-normal text-gray-500 dark:text-gray-400">
 							<div class="text-base font-semibold text-gray-900 dark:text-white">
-								{collector.name}
+								{source.name}
 							</div>
 						</div>
 					</TableBodyCell>
-					<TableBodyCell class="p-4">{collector.technology}</TableBodyCell>
+					<TableBodyCell class="p-4">{source.technology}</TableBodyCell>
 					<TableBodyCell
 						class="max-w-md overflow-hidden truncate p-4 text-base font-normal text-gray-500 dark:text-gray-400 xl:max-w-lg"
-						>{collector.description}</TableBodyCell
+						>{source.description}</TableBodyCell
 					>
 
 					<TableBodyCell class="flex justify-end space-x-2 p-4">
-						<Button size="sm" class="gap-2 px-3" on:click={() => toggle(Collector)}>
+						<Button size="sm" class="gap-2 px-3" on:click={() => toggle(Source)}>
 							<EditOutline size="sm" /> Add
 						</Button>
 					</TableBodyCell>
